+++ tests/test_joiner.py
@@ -31,3 +31,22 @@ def test_join_smart_url_with_path_and_basename():
     assert join_smart_url(["gsiftp://gridftp.zeuthen.desy.de:2811/pnfs/ifh.de/acs/icecube/archive/",
                            "/data/exp/IceCube/2015/filtered/level2/0320",
                            "fdd3c3865d1011eb97bb6224ddddaab7.zip"]) == CORRECT
+
+def test_join_smart_desy_bundle_path():
+    """Test join_smart functionality."""
+    CORRECT = "/pnfs/ifh.de/acs/icecube/archive/data/exp/IceCube/2015/filtered/level2/0320/604b6c80659c11eb8ad66224ddddaab7.zip"
+    assert join_smart(["/pnfs/ifh.de/acs/icecube/archive",
+                       "/data/exp/IceCube/2015/filtered/level2/0320",
+                       "604b6c80659c11eb8ad66224ddddaab7.zip"]) == CORRECT
+    assert join_smart(["/pnfs/ifh.de/acs/icecube/archive/",
+                       "/data/exp/IceCube/2015/filtered/level2/0320",
+                       "604b6c80659c11eb8ad66224ddddaab7.zip"]) == CORRECT
+    assert join_smart(["/pnfs/ifh.de/acs/icecube/archive",
+                       "/data/exp/IceCube/2015/filtered/level2/0320/",
+                       "604b6c80659c11eb8ad66224ddddaab7.zip"]) == CORRECT
+    assert join_smart(["/pnfs/ifh.de/acs/icecube/archive/",
+                       "/data/exp/IceCube/2015/filtered/level2/0320/",
+                       "604b6c80659c11eb8ad66224ddddaab7.zip"]) == CORRECT
+    assert join_smart(["/pnfs/ifh.de/acs/icecube/archive",
+                       "/data/exp/IceCube/2015/filtered/level2/0320",
+                       "/604b6c80659c11eb8ad66224ddddaab7.zip"]) == CORRECT